Various changes
Added PCB.C. Doesn't compile. Likely tons of pointer errors. InitPCBs() gets tons of errors. Apparently pcb1->chain = pcb2 is wrong. The manual says chain should be pointing to a physical location anyway so this needs to be changed.

Added the struct and definitions for PCBs in mpx.

Added a switch statement for "Show" command in comhan specifically for testing PCB.