Mod4 Finished
Finished Mod4. Basically the only relavent change is in MANAGECO.C lines 30 and 34. Build_PCB had the last two arguments wrong. segment has to be passed as the _CS and 0 as the IP.  IP is 0 because it's the pointer to the instructions in the _CS but it obviously has to start at 0. Otherwise everything else is basically unchanged.